refactor(cache): apply spec corrections to #190 re-key
- Rename migration to 014_lyrics_cache_duration_key.sql per spec
- Rewrite migration using the 012 NO TRANSACTION + explicit PRAGMA
  foreign_keys toggle + BEGIN/COMMIT pattern for crash safety
- Change durationBucket type from int64 to int throughout cache
  interface, worker.Cache interface, scan.LyricsCache interface,
  and all fake implementations in tests
- Remove the DurationBucket helper (floor(seconds/5) lands in #191;
  #190 only ever stores sentinel 0)
- Simplify cache tests to use integer literals instead of helper calls

+++ internal/scan/enqueuer.go
@@ -22,7 +22,7 @@ type PendingResultStore interface {
 type LyricsCache interface {
 	// Lookup checks the cache for (artist, title, durationBucket).
 	// Pass durationBucket=0 when the recording duration is not yet known.
-	Lookup(ctx context.Context, artist, title string, durationBucket int64) (string, error)
+	Lookup(ctx context.Context, artist, title string, durationBucket int) (string, error)
 }
 
 // WorkQueue enqueues durable lyrics work.
